functions/requirements.txt
# Firebase Admin
firebase-admin>=6.2.0
google-cloud-storage>=2.14.0
google-cloud-functions>=1.0.0

# Firebase Functions Python SDK
firebase-functions==0.1.0

# Google Cloud Storage
google-cloud-firestore>=2.16.0

# Flask app dependencies
flask==3.0.0
werkzeug==3.0.0
markdown==3.5.1
python-dateutil==2.8.2
python-dotenv==1.0.0
